feat: add user dropdown menu
