+++ frontend/sprint/src/scss/Notification.scss
@@ -13,14 +13,3 @@
     padding-bottom: 24px;
   }
 }
-
-.bell-button{
-  border: none;
-  box-shadow: none;
-  transform: translateY(2px);
-  overflow: hidden;
-}
-
-.bell-button:active{
-  border: none;
-}
